requirements.txt
python-twitter==3.3
requests==2.18.4
virtualenv==15.1.0
